pyproject.toml
[tool.poetry]
name = "vinbd"
version = "0.1.0"
description = ""
authors = ["BBOYADAO <bboyadao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
flakehell = "^0.9.0"
coverage = "^6.4.2"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
django-extensions = "^3.2.1"
daphne = "^3.0.2"


[tool.poetry.dev-dependencies]
ipython = "^8.7.0"
black = "^22.1.0"
flake8 = "^5.0.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
autoflake = "^1.4"
autoimport = "^1.2.2"
autopep8 = "^1.6.0"
reorder-python-imports = "^3.8.2"
yapf = "^0.32.0"
pylint = "^2.14.5"
django-extensions = "^3.2.0"
Werkzeug = "^2.1.2"

[tool.pyright]
include = ["./"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs"
]
ignore = ["src/oldstuff"]
stubPath = "./stubs"

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.10"
pythonPlatform = "Linux"

executionEnvironments = [
    # { root = "src/web", pythonVersion = "3.5", pythonPlatform = "Windows", extraPaths = [ "src/service_libs" ] },
    # { root = "src/sdk", pythonVersion = "3.0", extraPaths = [ "src/backend" ] },
    # { root = "src/tests", extraPaths = ["src/tests/e2e", "src/sdk" ]},
    { root = "./" }
]


[tool.coverage.paths]
source = [".", ]

[tool.coverage.run]
branch = true
source = ["."]
omit = ["*/tests/*",
    "*/migrations/*",
#    "*/urls.py",
#    "*/settings/*",
    "*/wsgi.py",
    "manage.py",
#    "*/apps.py",
#    "*/tasks.py",
#    "*/asgi.py"
]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# list of plugins and rules for them
[tool.flakehell.plugins]
# include everything in pyflakes except F401
pyflakes = ["+*", "-F401"]
# enable only codes from S100 to S199
flake8-bandit = ["-*", "+S1??"]
# enable everything that starts from `flake8-`
"flake8-*" = ["+*"]
# explicitly disable plugin
flake8-docstrings = ["-*"]
# 80 chars limit isn't enough in 21 century
max_line_length = 120
